chore(uat): move EventStorage back to control
